+++ src/widget/color_resolver.rs
@@ -1,5 +1,5 @@
 use crate::widget::EffectSpan;
-use crate::{RangeSampler, SimpleRng};
+use crate::{HslConvertable, RangeSampler, SimpleRng};
 use bon::builder;
 use ratatui::prelude::Color;
 use std::collections::BTreeSet;
@@ -42,7 +42,7 @@ impl ColorResolver {
         let len = effect_identifiers.len();
         let mut colors: Vec<Color> = (0..len)
             .map(|idx| hue.start + hue_range * idx as f64 / len as f64)
-            .map(|hue| Color::from_hsl(hue, saturation as _, lightness as _))
+            .map(|hue| Color::from_hsl(hue as _, saturation as _, lightness as _))
             .collect();
 
         let mut lcg = SimpleRng::default();